requirements.txt
streamlit>=1.33
pandas>=2.1
requests>=2.31
python-dotenv>=1.0
plotly>=5.20
pydantic>=2.5
numpy>=1.26
scipy>=1.11
httpx>=0.26
tenacity>=8.2
google-generativeai>=0.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
